requirements.txt
django
djangorestframework
celery
redis
python-dotenv
